# src/main/java/personal_projects/url_shortner/entity/Urls.java
package personal_projects.url_shortner.entity;

import org.springframework.data.annotation.Id;
import org.springframework.data.mongodb.core.mapping.Document;


@Document(collection = "urls")
public class Urls {

    private String key;
    private String value;



    public Urls(){};


    public Urls(String key, String value) {
        this.key = key;
        this.value = value;
    }

    public String getKey() {
        return key;
    }

    public void setKey(String key) {
        this.key = key;
    }

    public String getValue() {
        return value;
    }

    public void setValue(String value) {
        this.value = value;
    }
}